Shrink ints to gain some perf

+++ src/day16.rs
@@ -2,7 +2,7 @@ use std::iter;
 
 pub fn part1(file_data: &str)
 {
-    let mut signal = file_data.chars().map(|c| c.to_digit(10).unwrap() as i64 ).collect::<Vec<_>>();
+    let mut signal = file_data.chars().map(|c| c.to_digit(10).unwrap() as i32 ).collect::<Vec<_>>();
 
     for _ in 0..100
     {
@@ -17,14 +17,14 @@ pub fn part1(file_data: &str)
 
 pub fn part2(file_data: &str)
 {
-    let mut init_signal = file_data.chars().map(|c| c.to_digit(10).unwrap() as i64 ).cycle().take(file_data.len() * 10000).collect::<Vec<_>>();
+    let mut init_signal = file_data.chars().map(|c| c.to_digit(10).unwrap() as i32 ).cycle().take(file_data.len() * 10000).collect::<Vec<_>>();
     let skip_num = init_signal.iter().take(7).fold(0, |s, i| s*10 + i) as usize;
 
     let signal = &mut init_signal[skip_num..];
 
     for _ in 0..100
     {
-        let mut signal_sum = signal.iter().sum::<i64>();
+        let mut signal_sum = signal.iter().sum::<i32>();
         for i in 0..signal.len()
         {
             let cur_value = signal[i];
@@ -36,8 +36,8 @@ pub fn part2(file_data: &str)
     println!("Day 16 part 2: {:08}", signal.iter().take(8).fold(0, |s, i| s*10 + i));
 }
 
-fn calc_digit(digit: usize, signal: &Vec<i64>) -> i64
+fn calc_digit(digit: usize, signal: &Vec<i32>) -> i32
 {
     let cycle = [0,1,0,-1].iter().flat_map(|i| iter::repeat(i).take(digit)).cycle();
-    signal.iter().zip(cycle.skip(1)).map(|(x,y)| (x * y)).sum::<i64>().abs() % 10
+    signal.iter().zip(cycle.skip(1)).map(|(x,y)| (x * y)).sum::<i32>().abs() % 10
 }
